+++ lib/src/indicators/circle_stroke_spin.dart
@@ -9,6 +9,8 @@ class CircleStrokeSpin extends StatelessWidget {
     return CircularProgressIndicator(
       strokeWidth: 2,
       color: color,
+      backgroundColor:
+          DecorateContext.of(context)!.decorateData.pathBackgroundColor,
     );
   }
 }